Rewrote structures for D3D, added STB-Image for img loading, applied code style to entire solution, added gameobjects components

+++ amGraphicsDX/Source/amDXIndexBuffer.cpp
@@ -5,10 +5,12 @@ namespace amEngineSDK {
   amDXIndexBuffer::amDXIndexBuffer() {}
 
   amDXIndexBuffer::~amDXIndexBuffer() {}
-  void amDXIndexBuffer::setBufferData(D3D11_USAGE _usageF) {
+
+  void 
+  amDXIndexBuffer::setBufferData(D3D11_USAGE _usageF) {
     memset(&m_bd, 0, sizeof(m_bd));
     m_bd.Usage = _usageF;
-    m_bd.ByteWidth = sizeof(uint32) * m_vecIB.size();
+    m_bd.ByteWidth = static_cast<uint32>(sizeof(uint32) * m_vecIB.size());
     m_bd.BindFlags = D3D11_BIND_INDEX_BUFFER;
     m_bd.CPUAccessFlags = 0;
     m_bd.MiscFlags = 0;
@@ -19,15 +21,14 @@ namespace amEngineSDK {
     m_vecIB.clear();
   }
 
-  void amDXIndexBuffer::setData(Vector<uint32>& data) {
+  void 
+  amDXIndexBuffer::setData(Vector<uint32>& data) {
     m_vecIB = data;
   }
 
-  void amDXIndexBuffer::clear() {
+  void 
+  amDXIndexBuffer::clear() {
     m_vecIB.clear();
     m_vecIB.resize(0);
   }
 }
-
-
-
